tip: optimize label struct size

+++ source/widgets/label/sgl_label.h
@@ -43,7 +43,7 @@ typedef struct sgl_label {
     sgl_obj_t        obj;
     const sgl_font_t *font;
     char             *text;
-    size_t           text_capacity;
+    uint16_t         text_capacity;
     sgl_color_t      color;
     sgl_color_t      bg_color;
     union {
